[FIX/#81] 에러 다이얼로그 표시 시점 변경
